Bug 961751 - Split large test files into multiple smaller ones
To allow quickly running the tests for one view, for example.
