+++ mozillians/groups/tests/test_views_show.py
@@ -1,0 +1,272 @@
+from django.core.urlresolvers import reverse
+from django.test import Client
+from mozillians.common.tests import TestCase, requires_login, requires_vouch
+from mozillians.groups.models import GroupMembership
+from mozillians.groups.tests import GroupFactory, GroupAliasFactory, SkillFactory
+from mozillians.users.tests import UserFactory
+from nose.tools import eq_, ok_
+
+
+class ShowTests(TestCase):
+    def setUp(self):
+        self.group = GroupFactory.create()
+        self.url = reverse('groups:show_group', kwargs={'url': self.group.url})
+        self.user_1 = UserFactory.create(userprofile={'is_vouched': True})
+        self.user_2 = UserFactory.create(userprofile={'is_vouched': True})
+        self.group.add_member(self.user_2.userprofile)
+
+    def test_show_user_not_in_group(self):
+        with self.login(self.user_1) as client:
+            response = client.get(self.url, follow=True)
+        eq_(response.status_code, 200)
+        context = response.context
+        eq_(context['group'], self.group)
+        eq_(context['in_group'], False)
+        eq_(context['people'].paginator.count, 1)
+        eq_(context['people'][0], self.user_2.userprofile)
+        eq_(context['people'][0].pending, False)
+
+    def test_show_user_in_group(self):
+        """Test show() for a user within the group."""
+        with self.login(self.user_2) as client:
+            response = client.get(self.url, follow=True)
+        eq_(response.status_code, 200)
+        context = response.context
+        eq_(context['group'], self.group)
+        eq_(context['in_group'], True)
+        eq_(context['people'].paginator.count, 1)
+        eq_(context['people'][0], self.user_2.userprofile)
+
+    def test_show_pending_user(self):
+        self.group.add_member(self.user_2.userprofile, GroupMembership.PENDING)
+        with self.login(self.user_2) as client:
+            response = client.get(self.url, follow=True)
+        eq_(response.status_code, 200)
+        context = response.context
+        eq_(context['group'], self.group)
+        eq_(context['in_group'], False)
+        eq_(context['people'].paginator.count, 1)
+        eq_(context['people'][0], self.user_2.userprofile)
+        eq_(context['people'][0].pending, True)
+
+    def test_show_empty_group(self):
+        group = GroupFactory.create()
+        url = reverse('groups:show_group', kwargs={'url': group.url})
+        with self.login(self.user_1) as client:
+            response = client.get(url, follow=True)
+        eq_(response.status_code, 200)
+        context = response.context
+        eq_(context['people'].paginator.count, 0)
+
+    @requires_login()
+    def test_show_anonymous(self):
+        client = Client()
+        client.get(self.url, follow=True)
+
+    @requires_vouch()
+    def test_show_unvouched(self):
+        user = UserFactory.create()
+        with self.login(user) as client:
+            client.get(self.url, follow=True)
+
+    def test_nonexistant_group(self):
+        url = reverse('groups:show_group', kwargs={'url': 'invalid'})
+        with self.login(self.user_1) as client:
+            response = client.get(url, follow=True)
+        eq_(response.status_code, 404)
+
+    def test_alias_redirection(self):
+        user = UserFactory.create(userprofile={'is_vouched': True})
+        group = GroupFactory.create()
+        group_alias = GroupAliasFactory.create(alias=group)
+        url = reverse('groups:show_group', kwargs={'url': group_alias.url})
+        with self.login(user) as client:
+            response = client.get(url, follow=True)
+        eq_(response.status_code, 200)
+        eq_(response.context['group'], group)
+
+    def test_show_leave_button_value_with_curator(self):
+        curator_user = UserFactory.create(userprofile={'is_vouched': True})
+        group = GroupFactory.create(curator=curator_user.userprofile)
+        user = UserFactory.create(userprofile={'is_vouched': True})
+        group.add_member(user.userprofile)
+        url = reverse('groups:show_group', kwargs={'url': group.url})
+
+        with self.login(curator_user) as client:
+            response = client.get(url, follow=True)
+        eq_(response.status_code, 200)
+        eq_(response.context['show_leave_button'], False)
+
+        with self.login(user) as client:
+            response = client.get(url, follow=True)
+        eq_(response.status_code, 200)
+        eq_(response.context['show_leave_button'], True)
+
+    def test_show_leave_button_value_without_curator(self):
+        group = GroupFactory.create()
+        user = UserFactory.create(userprofile={'is_vouched': True})
+        group.add_member(user.userprofile)
+        url = reverse('groups:show_group', kwargs={'url': group.url})
+
+        with self.login(user) as client:
+            response = client.get(url, follow=True)
+        eq_(response.status_code, 200)
+        eq_(response.context['show_leave_button'], True)
+
+    def test_show_leave_button_value_members_cant_leave(self):
+        """
+        Don't show leave button for a group whose members_can_leave flag
+        is False, even for group member
+        """
+        group = GroupFactory.create(members_can_leave=False)
+        user = UserFactory.create(userprofile={'is_vouched': True})
+        group.add_member(user.userprofile)
+        url = reverse('groups:show_group', kwargs={'url': group.url})
+
+        with self.login(user) as client:
+            response = client.get(url, follow=True)
+        eq_(response.status_code, 200)
+        eq_(response.context['show_leave_button'], False)
+
+    def test_show_leave_button_value_members_can_leave(self):
+        """
+        Do show leave button for a group whose members_can_leave flag
+        is True, for group member
+        """
+        group = GroupFactory.create(members_can_leave=True)
+        user = UserFactory.create(userprofile={'is_vouched': True})
+        group.add_member(user.userprofile)
+        url = reverse('groups:show_group', kwargs={'url': group.url})
+
+        with self.login(user) as client:
+            response = client.get(url, follow=True)
+        eq_(response.status_code, 200)
+        eq_(response.context['show_leave_button'], True)
+
+    def test_show_leave_button_value_members_can_leave_non_member(self):
+        """
+        Don't show leave button for a group whose members_can_leave flag
+        is True, if not group member
+        """
+        group = GroupFactory.create(members_can_leave=True)
+        user = UserFactory.create(userprofile={'is_vouched': True})
+        url = reverse('groups:show_group', kwargs={'url': group.url})
+
+        with self.login(user) as client:
+            response = client.get(url, follow=True)
+        eq_(response.status_code, 200)
+        eq_(response.context['show_leave_button'], False)
+
+    def test_show_join_button_accepting_members_yes(self):
+        group = GroupFactory.create(accepting_new_members='yes')
+        user = UserFactory.create(userprofile={'is_vouched': True})
+        url = reverse('groups:show_group', kwargs={'url': group.url})
+
+        with self.login(user) as client:
+            response = client.get(url, follow=True)
+        eq_(response.status_code, 200)
+        eq_(response.context['show_join_button'], True)
+
+    def test_show_join_button_accepting_members_yes_member(self):
+        group = GroupFactory.create(accepting_new_members='yes')
+        user = UserFactory.create(userprofile={'is_vouched': True})
+        group.add_member(user.userprofile)
+        url = reverse('groups:show_group', kwargs={'url': group.url})
+
+        with self.login(user) as client:
+            response = client.get(url, follow=True)
+        eq_(response.status_code, 200)
+        eq_(response.context['show_join_button'], False)
+
+    def test_show_join_button_accepting_members_by_request(self):
+        group = GroupFactory.create(accepting_new_members='yes')
+        user = UserFactory.create(userprofile={'is_vouched': True})
+        url = reverse('groups:show_group', kwargs={'url': group.url})
+
+        with self.login(user) as client:
+            response = client.get(url, follow=True)
+        eq_(response.status_code, 200)
+        eq_(response.context['show_join_button'], True)
+
+    def test_show_join_button_accepting_members_by_request_member(self):
+        group = GroupFactory.create(accepting_new_members='yes')
+        user = UserFactory.create(userprofile={'is_vouched': True})
+        group.add_member(user.userprofile)
+        url = reverse('groups:show_group', kwargs={'url': group.url})
+
+        with self.login(user) as client:
+            response = client.get(url, follow=True)
+        eq_(response.status_code, 200)
+        eq_(response.context['show_join_button'], False)
+
+    def test_show_join_button_accepting_members_no(self):
+        group = GroupFactory.create(accepting_new_members='no')
+        user = UserFactory.create(userprofile={'is_vouched': True})
+        url = reverse('groups:show_group', kwargs={'url': group.url})
+
+        with self.login(user) as client:
+            response = client.get(url, follow=True)
+        eq_(response.status_code, 200)
+        eq_(response.context['show_join_button'], False)
+
+    def test_show_leave_button_value_skill(self):
+        skill = SkillFactory.create()
+        user = UserFactory.create(userprofile={'is_vouched': True})
+        skill.members.add(user.userprofile)
+        url = reverse('groups:show_skill', kwargs={'url': skill.url})
+
+        with self.login(user) as client:
+            response = client.get(url, follow=True)
+        eq_(response.status_code, 200)
+        eq_(response.context['show_leave_button'], True)
+
+    def test_remove_button_confirms(self):
+        """GET to remove_member view displays confirmation"""
+        # Make user 1 the group curator so they can remove users
+        self.group.curator = self.user_1.userprofile
+        self.group.save()
+
+        # We must request the full path, with language, or the
+        # LanguageMiddleware will convert the request to GET.
+        url = reverse('groups:remove_member', prefix='/en-US/',
+                      kwargs=dict(group_pk=self.group.pk, user_pk=self.user_2.userprofile.pk))
+        with self.login(self.user_1) as client:
+            response = client.get(url, follow=True)
+        self.assertTemplateUsed(response, 'groups/confirm_remove_member.html')
+        # Still a member
+        ok_(self.group.has_member(self.user_2.userprofile))
+
+    def test_post_remove_button_removes(self):
+        """POST to remove_member view removes member"""
+        # Make user 1 the group curator so they can remove users
+        self.group.curator = self.user_1.userprofile
+        self.group.save()
+
+        # We must request the full path, with language, or the
+        # LanguageMiddleware will convert the request to GET.
+        url = reverse('groups:remove_member', prefix='/en-US/',
+                      kwargs=dict(group_pk=self.group.pk, user_pk=self.user_2.userprofile.pk))
+        with self.login(self.user_1) as client:
+            response = client.post(url, follow=True)
+        self.assertTemplateNotUsed(response, 'groups/confirm_remove_member.html')
+        # Not a member anymore
+        ok_(not self.group.has_member(self.user_2.userprofile))
+
+    def test_confirm_user(self):
+        """POST to confirm user view changes member from pending to member"""
+        # Make user 1 the group curator so they can remove users
+        self.group.curator = self.user_1.userprofile
+        self.group.save()
+        # Make user 2 pending
+        self.group.add_member(self.user_2.userprofile, GroupMembership.PENDING)
+        ok_(self.group.has_pending_member(self.user_2.userprofile))
+
+        # We must request the full path, with language, or the
+        # LanguageMiddleware will convert the request to GET.
+        url = reverse('groups:confirm_member', prefix='/en-US/',
+                      kwargs=dict(group_pk=self.group.pk, user_pk=self.user_2.userprofile.pk))
+        with self.login(self.user_1) as client:
+            response = client.post(url, follow=True)
+        self.assertTemplateNotUsed(response, 'groups/confirm_remove_member.html')
+        # Now a member
+        ok_(self.group.has_member(self.user_2.userprofile))
